+++ spotbugsTestCases/src/java/increasedAccessibilityOfMethods/Interface.java
@@ -1,0 +1,7 @@
+package increasedAccessibilityOfMethods;
+
+public interface Interface {
+
+    String packagePrivateInterfaceMethod();
+
+}